django/learnyBackEnd/requirements.txt
Package             Version
------------------- ---------
asgiref             3.6.0
asttokens           2.2.1
backcall            0.2.0
certifi             2022.12.7
colorama            0.4.6
comm                0.1.2
debugpy             1.6.5
decorator           5.1.1
Django              4.1.5
django-extensions   3.2.1
djangorestframework 3.14.0
entrypoints         0.4
executing           1.2.0
importlib-metadata  6.0.0
ipykernel           6.20.1
ipython             8.8.0
jedi                0.18.2
jupyter_client      7.4.9
jupyter_core        5.1.3
Markdown            3.4.1
matplotlib-inline   0.1.6
nest-asyncio        1.5.6
numpy               1.24.1
packaging           23.0
pandas              1.5.2
parso               0.8.3
pickleshare         0.7.5
pip                 22.3.1
platformdirs        2.6.2
prompt-toolkit      3.0.36
psutil              5.9.4
pure-eval           0.2.2
Pygments            2.14.0
python-dateutil     2.8.2
pytz                2022.7.1
pywin32             305
pyzmq               25.0.0
setuptools          65.6.3
six                 1.16.0
sqlparse            0.4.3
stack-data          0.6.2
tornado             6.2
traitlets           5.8.1
tzdata              2022.7
wcwidth             0.2.6
wheel               0.37.1
wincertstore        0.2
zipp                3.11.0
